Extracting chat types to separate file
