+++ src/chat/redux/index.ts
@@ -1,3 +1,4 @@
 export * from './chatActions';
 export * from './chatReducer';
 export * from './chatSagas';
+export * from './types';
